Adding step by step descriptions for adding email channels
This commit updates the email channel creation by a step by step description.

This hopefully helps administrators to spot possible issues during configuration, see scopes and understand the specific settings better.
It hopefully brings in more light in some magic boxes of Zammad

Also, it adds information on how to configure outbound  as this earlier was completely ignored in the docs. *whoops*